Fixed the delete imgae issue

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -185,10 +185,28 @@ public class ImageController {
 	// The method calls the deleteImage() method in the business logic passing the
 	// id of the image to be deleted
 	// Looks for a controller method with request mapping of type '/images'
-	@RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-	public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-		imageService.deleteImage(imageId);
-		return "redirect:/images";
+	@RequestMapping(value = "/deleteImage", method = RequestMethod.POST)
+	public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session,Model model) {
+		
+		Image image = imageService.getImage(imageId);
+		
+		User loggeduser = (User) session.getAttribute("loggeduser");
+		User imgeOwner = image.getUser();
+
+		String tags = convertTagsToString(image.getTags());
+		model.addAttribute("image", image);
+		model.addAttribute("tags", tags);
+		
+		if (loggeduser.getUsername().equalsIgnoreCase(imgeOwner.getUsername())) {
+			imageService.deleteImage(imageId);
+			return "redirect:/images";
+		} else {
+			String error = "Only the owner of the image can delete the image";
+			model.addAttribute("deleteError", error);
+			return "images/image";
+		}
+		
+		
 	}
 
 	// This method converts the image to Base64 format
